feat: add BigInt arithmetic support

+++ src/main/kotlin/calculator/OutputView.kt
@@ -1,7 +1,9 @@
 package calculator
 
+import java.math.BigInteger
+
 class OutputView {
-    fun printResult(result: Int) {
+    fun printResult(result: BigInteger) {
         println("결과 : $result")
     }
 }
